requirements.txt
# No external dependencies required for this simple application 
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
deepface>=0.0.79
openai>=1.76.0
python-dotenv>=1.0.0
pillow>=10.0.0 